#81829  - remove io directory, move function to deprecated functions.php, remove calls to deprecated options methods
