+++ src/Tribe/Importer/File_Importer_Events.php
@@ -150,9 +150,7 @@ class Tribe__Events__Importer__File_Importer_Events extends Tribe__Events__Impor
 		$start_date = strtotime( $this->get_event_start_date( $record ) );
 		$end_date   = strtotime( $this->get_event_end_date( $record ) );
 
-		if ( empty( $this->is_aggregator ) ) {
-			$post_status_setting = Tribe__Events__Importer__Options::get_default_post_status( 'csv' );
-		} elseif ( $this->default_post_status ) {
+		if ( $this->default_post_status ) {
 			$post_status_setting = $this->default_post_status;
 		} else {
 			$post_status_setting = tribe( 'events-aggregator.settings' )->default_post_status( 'csv' );
