+++ src/Tribe/Main.php
@@ -470,10 +470,6 @@ if ( ! class_exists( 'Tribe__Events__Main' ) ) {
 			// Tribe common resources
 			require_once $this->plugin_path . 'vendor/tribe-common-libraries/tribe-common-libraries.class.php';
 
-			// Load CSV importer
-			//todo deprecate
-			//require_once $this->plugin_path . 'src/io/csv/ecp-events-importer.php';
-
 			// Load Template Tags
 			require_once $this->plugin_path . 'src/functions/template-tags/query.php';
 			require_once $this->plugin_path . 'src/functions/template-tags/general.php';
