--- src/io/csv/admin-views/footer.php
@@ -1,7 +1,0 @@
-<?php
-// Don't load directly
-if ( ! defined( 'ABSPATH' ) ) {
-	die( '-1' );
-}
-?>
-</div></div>
